misc/HtmlTestSuite/results/test-0060.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head><meta charset="utf-8"/>
<title>Event Timing: supportedEntryTypes should be consistent with `'PerformanceEventTiming' in Self`
</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script>
test(() =&gt; {
  const isFirstInputSupportedInEntryTypes = PerformanceObserver.supportedEntryTypes.includes('first-input');
  const isEventSupportedInEntryTypes = PerformanceObserver.supportedEntryTypes.includes('event');

  const isEventTimingSupportedInSelf = 'PerformanceEventTiming' in self;

  assert_equals(isFirstInputSupportedInEntryTypes, isEventTimingSupportedInSelf)
  assert_equals(isEventSupportedInEntryTypes, isEventTimingSupportedInSelf)
}, "supportedEntryTypes should be consistent with `'PerformanceEventTiming' in Self`");
</script>
</head><body/></html>
